moved /sections/category/{id} to category/{id}/section
